4.0 ALPHA THE BIG UPDATE
So i am aware that there where many bugs, most of them caused by the
clutter of all the code in one class. So i decided to organize a bit,
and recode everything. (I kept most big parts that worked fine)

So this is ALPHA. This is 4.0 with the new code, But is at the same
working state as the one now (even with the same logic bugs) the BETA
update will fix all of these.