Make usage of Price more comprehensive
There will now be a PRICETABLE entry for every item in the shop. If the
item does not have a buy-price, Price#getBuyPrice == -1. If the item
does not have a sell-price, Price#getSellPrice == -1. Also, when loading
items and inserting Price objects to the pricetable, use computeIfAbsent
to avoid unnecessary churn.